Done : moderating list of polls and pending polls
Data is being populated from socket
Questions that do not need to wait for resolving are checked on the spot while the user is wagering, and they are not added to the pending list. After the answer is checked, that poll is removed from the list.
Questions that need to wait for resolving are added to the waiting list. When their associated resolve message arrives, the win or lost message is displayed, and the question is removed from the waiting and poll lists.
The current implementation is in a test app, so it was not possible to record a clip. It will be set up in the actual free app tomorrow.